Centralized usage of google library
